+++ src/lib-public/FA2-generator.ts
@@ -58,6 +58,7 @@ export function generateFA2(invoice: Faktura, additionalData: AdditionalDataType
         ...generateStopka(additionalData, invoice.Stopka, invoice.Naglowek, invoice.Fa?.WZ),
       ];
     const docDefinition: TDocumentDefinitions = {
+      watermark: additionalData?.watermark,
       content,
       footer: (currentPage, pageCount) => {
         return {