+++ src/shared/PDF-functions.ts
@@ -19,7 +19,7 @@ import {
 } from './consts/FA.const';
 import { Kraj } from './consts/const';
 import { TStawkaPodatku_FARR } from './consts/FARR.const';
-import { formatDateTime, formatTime, translateMap } from './generators/common/functions';
+import { formatDateTimePl, formatTime, translateMap } from './generators/common/functions';
 import { HeaderDefine, PdfFP, PdfOptionField } from './types/pdf-types';
 import { FP } from '../lib-public/types/fa3.types';
 import { DifferentValues, FilteredKeysOfValues, TypesOfValues } from './types/universal.types';
@@ -148,10 +148,10 @@ function formatValue(
       result.alignment = Position.RIGHT;
       break;
     case FormatTyp.DateTime:
-      result.text = formatDateTime(value as string);
+      result.text = formatDateTimePl(value as string, true, true);
       break;
     case FormatTyp.Date:
-      result.text = formatDateTime(value as string, false, true);
+      result.text = formatDateTimePl(value as string);
       break;
     case FormatTyp.Time:
       result.text = formatTime(value as string);
@@ -536,7 +536,12 @@ export function getContentTable<T>(
   };
 }
 
-export function generateTwoColumns(kol1: Column, kol2: Column, margin?: Margins): Content {
+export function generateTwoColumns(
+  kol1: Column,
+  kol2: Column,
+  margin?: Margins,
+  unbreakable = true
+): Content {
   return {
     columns: [
       { stack: [kol1], width: '50%' },
@@ -544,6 +549,7 @@ export function generateTwoColumns(kol1: Column, kol2: Column, margin?: Margins)
     ],
     margin: margin ?? [0, 0, 0, 0],
     columnGap: 20,
+    unbreakable,
   };
 }